major update 2.0
